Refactor setup script to conditionally install Miniconda

+++ setup.sh
@@ -3,9 +3,11 @@ set -e
 
 apt-get update && apt-get install -y git wget build-essential unzip
 
-wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
-bash miniconda.sh -b -p /opt/miniconda
-rm miniconda.sh
+if [ ! -d "/opt/miniconda" ]; then
+    wget https://repo.anaconda.com/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
+    bash miniconda.sh -b -p /opt/miniconda
+    rm miniconda.sh
+fi
 export PATH="/opt/miniconda/bin:$PATH"
 
 conda init bash
